Separate game board logic into distinct modules
