+++ game.html
@@ -90,26 +90,26 @@
           <div class="row px-2 pt-3 pb-4 player-stats-section">
             <div class="col">
               <div class="progress">
-                <div id="player-hp" class="progress-bar bg-warning" role="progressbar" aria-valuenow="85"
-                  aria-valuemin="0" aria-valuemax="100" style="width: 85%">
+                <div id="player-hp" class="progress-bar bg-warning" role="progressbar" aria-valuenow="50"
+                  aria-valuemin="0" aria-valuemax="100" style="width: 50%">
                   <span class="text-left pl-2 progress-item">HP</span>
                 </div>
               </div>
               <div class="progress">
-                <div id="player-strength" class="progress-bar bg-danger" role="progressbar" aria-valuenow="85"
-                  aria-valuemin="0" aria-valuemax="100" style="width: 85%">
+                <div id="player-strength" class="progress-bar bg-danger" role="progressbar" aria-valuenow="50"
+                  aria-valuemin="0" aria-valuemax="100" style="width: 50%">
                   <span class="text-left pl-2 progress-item">Strength</span>
                 </div>
               </div>
               <div class="progress">
-                <div id="player-agility" class="progress-bar bg-success" role="progressbar" aria-valuenow="70"
-                  aria-valuemin="0" aria-valuemax="100" style="width: 70%">
+                <div id="player-agility" class="progress-bar bg-success" role="progressbar" aria-valuenow="50"
+                  aria-valuemin="0" aria-valuemax="100" style="width: 50%">
                   <span class="text-left pl-2 progress-item">Agility</span>
                 </div>
               </div>
               <div class="progress">
-                <div id="player-dexterity" class="progress-bar bg-info" role="progressbar" aria-valuenow="80"
-                  aria-valuemin="0" aria-valuemax="100" style="width: 80%">
+                <div id="player-dexterity" class="progress-bar bg-info" role="progressbar" aria-valuenow="50"
+                  aria-valuemin="0" aria-valuemax="100" style="width: 50%">
                   <span class="text-left pl-2 progress-item">Dexterity</span>
                 </div>
               </div>
@@ -165,7 +165,9 @@
   <script src="https://cdn.jsdelivr.net/npm/interactjs/dist/interact.min.js"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/gsap.min.js"></script>
   <!-- Main script JS -->
-  <script type="module" src="assets/js/script.js"></script>
+  <script src="assets/js/script.js"></script>
+  <script defer type="module" src="assets/js/game_board.js"></script>
+  <script defer type="module" src="assets/js/potions_draggability.js"></script>
   <!-- Fontawesome Script -->
   <script src="https://kit.fontawesome.com/4e8d2c2bb0.js" crossorigin="anonymous"></script>
